Added options for modes
One can require the system to work in the regular mode(stablize to a setpoint) by making mode[i]=1, or work in the max-mode(rotate to highest power) by making mode[i]=0.
If you want two modes to be enabled at the same time, which means that the system will hop to max-mode when there's not enough power, set modechage[i]=1. To disable mode chage set modechange[i]=0.